style(subscriber): put aggregator in its own mod
